# internal/gateway/api/openapi.yaml
openapi: 3.0.1
info:
  title: Orochi object API
  version: 0.1.0
  description: API for orochi object manipulation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

security: []
paths:
  /{bucket}/{object}:
    put:
      summary: Create a object
      operationId: createObject
      parameters:
        - $ref: '#/components/parameters/bucket'
        - $ref: '#/components/parameters/object'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Object created successfully
        '400':
          description: Bucket should be created first

    get:
      summary: Get a object
      operationId: getObject
      parameters:
        - $ref: '#/components/parameters/bucket'
        - $ref: '#/components/parameters/object'
      responses:
        '200':
          description: Object received successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Bucket should be created first
        '404':
          description: Object not found

    delete:
      summary: Delete a object
      operationId: deleteObject
      parameters:
        - $ref: '#/components/parameters/bucket'
        - $ref: '#/components/parameters/object'
      responses:
        '204':
          description: Object deleted successfully
        '400':
          description: Bucket should be created first

  /{bucket}:
    get:
      summary: List objects
      operationId: listObjects
      parameters:
        - $ref: '#/components/parameters/bucket'
        - name: X-First-Object-ID
          in: header
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: X-Limit
          in: header
          description: The limit of the result array length
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Objects listed successfully
          headers:
            X-Next-Object-ID:
              description: |
                Next object ID for the continuous requests.
                If the last object is included in the result,
                this header is set to the maximum value of int64.
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/objectList'
        '400':
          description: Bucket should be created first

components:
  schemas:
    objectList:
      type: array
      items:
        type: string
  parameters:
    bucket:
      name: bucket
      in: path
      required: true
      schema:
        type: string
    object:
      name: object
      in: path
      required: true
      schema:
        type: string